move files based on section
